+++ Build1/zmachine.Library.Tests/TestZMachine.Opcodes.cs
@@ -11,60 +11,6 @@
     public class TestZMachineOpcodes
     {
         [TestMethod]
-        public void Test_pcGetByte()
-        {
-            IIO staticIO = new NullIO();
-            Machine machine = new Machine(
-                staticIO,
-                TestZMachine.ZorkPath,
-                new Dictionary<BreakpointType, BreakpointAction>
-                {
-                });
-            byte pcByte = machine.pc_getByte();
-        }
-
-        [TestMethod]
-        public void Test_setVar()
-        {
-            IIO staticIO = new NullIO();
-            Machine machine = new Machine(
-                staticIO,
-                TestZMachine.ZorkPath,
-                new Dictionary<BreakpointType, BreakpointAction>
-                {
-                });
-
-            ushort value = 2345;
-            ushort[] variables = new ushort[] { 0, 0x0f, 0xff };
-            foreach (ushort variable in variables)
-            {
-                machine.setVar(
-                    variable: variable,
-                    value: value);
-            }
-            // test all three branches of setVar
-            // make sure memory setWord was called
-            // Variable number $00 refers to the top of the stack
-            //$01 to $0f mean the local variables of the current routine
-            //and $10 to $ff mean the global variables.
-        }
-
-        [TestMethod]
-        public void Test_popRoutineData()
-        {
-            IIO staticIO = new NullIO();
-            Machine machine = new Machine(
-                staticIO,
-                TestZMachine.ZorkPath,
-                new Dictionary<BreakpointType, BreakpointAction>
-                {
-                });
-
-            ushort returnVal = 12345;
-            machine.popRoutineData(returnValue: returnVal);
-        }
-
-        [TestMethod]
         public void Test_op_rtrue()
         {
             IIO staticIO = new NullIO();
